refactor: 로또 당첨 통계 관련 로직을 Lotto 클래스로 분리
