+++ src/Lotto.js
@@ -1,4 +1,4 @@
-const { ERROR } = require('./utils/constants');
+const { ERROR, WINNING_MONEY } = require('./utils/constants');
 class Lotto {
   #numbers;
 
@@ -23,6 +23,42 @@ class Lotto {
     if (Math.min(...numbers) <= 0 || Math.max(...numbers) > 45) {
       throw new Error(ERROR.OUT_OF_RANGE);
     }
+  }
+
+  getResult(lottoBundle, bonus) {
+    const result = new Array(6).fill(0);
+    lottoBundle.map((lotto) => {
+      const ranking = this.getWinningRanking(lotto, bonus);
+      result[ranking] += 1;
+    });
+    return result.slice(1, 6).reverse();
+  }
+
+  getWinningRanking(lotto, bonus) {
+    const matchingCount = lotto.filter((number) => this.#numbers.includes(number)).length;
+    const hasBonus = lotto.includes(bonus);
+    if (matchingCount === 6) {
+      return 1;
+    }
+    if (matchingCount === 5) {
+      if (hasBonus) return 2;
+      return 3;
+    }
+    if (matchingCount === 4) {
+      return 4;
+    }
+    if (matchingCount === 3) {
+      return 5;
+    }
+    return 0;
+  }
+
+  calculateProfitRate(result, money) {
+    const profit = result.reduce((acc, cur, idx) => {
+      return acc + WINNING_MONEY[idx] * cur;
+    }, 0);
+    const profitRate = (profit / money) * 100;
+    return profitRate.toFixed(1);
   }
 }
